Updated graphics systems
GraphicsUtility now has better error checking methods

+++ src/ritzow/solomon/engine/world/component/Luminous.java
@@ -1,9 +1,9 @@
 package ritzow.solomon.engine.world.component;
 
 public interface Luminous {
-	public short getLightRed();
-	public short getLightGreen();
-	public short getLightBlue();
+	public float getLightRed();
+	public float getLightGreen();
+	public float getLightBlue();
 	public float getLightRadius();
 	public float getLightIntensity();
 }
